rename: CursorPageResponse 삭제 및 위치 변경, 그에 따른 타 파일의 변수타입 및 import문 변경
autosync.dto의 위치에 있던 CursorPageResponse를 작업 연동 조회에서도 사용하게 되어 삭제
->
common.dto에 있던 PageSyncResponse를 공통으로 사용하도록 하고 이름을 CursorPageResponse로 변경
